commintment page that signs private data adds an index page with links
