+++ apps/frontend/components/commitment/CommitmentForm.tsx
@@ -6,7 +6,7 @@ import {
   FormLabel,
   Input,
 } from "@chakra-ui/react"
-import { useEffect } from "react"
+import { useMemo } from "react"
 import { useForm } from "react-hook-form"
 import { useAccount } from "wagmi"
 
@@ -16,9 +16,8 @@ export const CommitmentForm = (props: { onSubmit: (data: any) => void }) => {
   const {
     register,
     handleSubmit,
-    formState: { errors, isValid },
-    setValue,
-    control,
+    formState: { isValid },
+    watch,
   } = useForm({
     defaultValues: {
       privateAddress: address,
@@ -26,10 +25,10 @@ export const CommitmentForm = (props: { onSubmit: (data: any) => void }) => {
     },
     mode: "all",
   })
-
-  useEffect(() => {
-    setValue("privateAddress", address)
-  }, [address, setValue])
+  const watchFields = watch(["privateAddress", "anonAddress"])
+  const canSubmit = useMemo(() => {
+    return address !== watchFields[0] && address === watchFields[1]
+  }, [watchFields, address])
 
   return (
     <Flex
@@ -48,7 +47,6 @@ export const CommitmentForm = (props: { onSubmit: (data: any) => void }) => {
             required: true,
             minLength: 40,
           })}
-          disabled
         />
       </FormControl>
       <FormControl>
@@ -61,8 +59,8 @@ export const CommitmentForm = (props: { onSubmit: (data: any) => void }) => {
           })}
         />
       </FormControl>
-      <Button type="submit" disabled={!isValid}>
-        Submit
+      <Button type="submit" disabled={!isValid || !canSubmit}>
+        Hash & Sign
       </Button>
     </Flex>
   )
