+++ apps/frontend/pages/requests.tsx
@@ -1,4 +1,15 @@
-import { Heading, Table, Td, Th, Thead, Tr } from "@chakra-ui/react"
+import {
+  Button,
+  Heading,
+  Table,
+  Link as ChLink,
+  Td,
+  Text,
+  Th,
+  Thead,
+  Tr,
+} from "@chakra-ui/react"
+import Link from "next/link"
 import { useEffect, useState } from "react"
 import { useTokenContract } from "../hooks/useTokenContract"
 
@@ -26,15 +37,29 @@ export default function RequestList() {
 
           <Th>Deadline</Th>
           <Th>Paper CID</Th>
+          <Th>Actions</Th>
         </Tr>
         {allReviewRequests.map((r) => {
           return (
             <Tr key={r.args.antId}>
               <Td>{r.args.antId.toNumber()}</Td>
-              <Td>{r.args.issuers[0]}</Td>
+              <Td>
+                <Text maxWidth={100} isTruncated>
+                  {r.args.issuers[0]}
+                </Text>
+              </Td>
 
               <Td>{new Date(r.args.deadline.toNumber()).toISOString()}</Td>
-              <Td>{r.args.paperHash}</Td>
+              <Td>
+                <Text maxWidth={100} isTruncated>
+                  {r.args.paperHash}
+                </Text>
+              </Td>
+              <Td>
+                <Link href={`/commit/${r.args.antId}`} passHref legacyBehavior>
+                  <ChLink as={Button}>Commit</ChLink>
+                </Link>
+              </Td>
             </Tr>
           )
         })}
